requirements.txt
torch
torchvision
opencv-python
pandas
matplotlib
streamlit
PyYAML
tqdm
